+++ src/Resources/views/livewire/pages/permissions.blade.php
@@ -1,4 +1,8 @@
-@extends('permission-wrapper::layouts.app')
+@php
+$namespace = config('permissions-ui.views.namespace', 'permission-wrapper');
+@endphp
+
+@extends($namespace . '::layouts.app')
 
 @section('title', 'Permissions Management')