+++ src/Resources/views/tailwind/create.blade.php
@@ -1,4 +1,8 @@
-@extends('permissions-ui::layouts.app')
+@php
+$namespace = config('permissions-ui.views.namespace', 'permission-wrapper');
+@endphp
+
+@extends($namespace . '::layouts.app')
 
 @section('content')
 <div class="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">